# spec/swagger.yaml
definitions:
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.SkillItem:
    properties:
      content:
        type: string
      description:
        type: string
      managed:
        type: boolean
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
      raw:
        type: string
      shadowed_by:
        type: string
      source_kind:
        type: string
      source_path:
        type: string
      source_root:
        type: string
      state:
        type: string
    type: object
  handlers.SkillsActionRequest:
    properties:
      action:
        type: string
      target_path:
        type: string
    type: object
  handlers.SkillsDeleteRequest:
    properties:
      names:
        items:
          type: string
        type: array
    type: object
  handlers.SkillsResponse:
    properties:
      skills:
        items:
          $ref: '#/definitions/handlers.SkillItem'
        type: array
    type: object
  handlers.SkillsUpsertRequest:
    properties:
      skills:
        items:
          type: string
        type: array
    type: object
...
  handlers.SupermarketAuthor:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
...
  handlers.SupermarketSkillEntry:
    properties:
      content:
        type: string
      description:
        type: string
      files:
        items:
          type: string
        type: array
      id:
        type: string
      metadata:
        $ref: '#/definitions/handlers.SupermarketSkillMetadata'
      name:
        type: string
    type: object
  handlers.SupermarketSkillListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.SupermarketSkillEntry'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  handlers.SupermarketSkillMetadata:
    properties:
      author:
        $ref: '#/definitions/handlers.SupermarketAuthor'
      homepage:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
  handlers.skillsOpResponse:
    properties:
      ok:
        type: boolean
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /bots/{bot_id}/container/skills:
    delete:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Delete skills payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.SkillsDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.skillsOpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete Memoh-managed skills
      tags:
      - workspace
    get:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SkillsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List skills from the bot container
      tags:
      - workspace
    post:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Skills payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.SkillsUpsertRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.skillsOpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Upload skills into Memoh-managed directory
      tags:
      - workspace
  /bots/{bot_id}/container/skills/actions:
    post:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Skill action payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.SkillsActionRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.skillsOpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Apply an action to a discovered or managed skill source
      tags:
      - workspace
...
  /bots/{bot_id}/tools:
    post:
      description: MCP endpoint for tool discovery and invocation.
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: JSON-RPC request
        in: body
        name: payload
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 'JSON-RPC response: {jsonrpc,id,result|error}'
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Unified MCP tools gateway
      tags:
      - workspace
...
  /supermarket/skills:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SupermarketSkillListResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List skills from supermarket
      tags:
      - supermarket
  /supermarket/skills/{id}:
    get:
      parameters:
      - description: Skill ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SupermarketSkillEntry'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get skill detail from supermarket
      tags:
      - supermarket
...
swagger: "2.0"